Fixed: AndroidManifest.xml definitions Added: ReturnSet and ErrorCode classes Fixed: File Exists checks

+++ Src/Bro2Bro.Mobile/Bro2Bro.Mobile.Droid/Interfaces/LocalStorage.cs
@@ -4,9 +4,14 @@ using System.Text;
 
 using Newtonsoft.Json;
 
-using Bro2Bro.Mobile.Interfaces;
+using Xamarin.Forms;
 
-namespace Bro2Bro.Mobile.Droid.Interfaces {
+using Bro2Bro.Mobile.Interfaces;
+using Bro2Bro.PCL.Transports.Global;
+using Bro2Bro.Mobile.Droid.Interfaces;
+
+[assembly: Dependency(typeof(LocalStorage))]
+namespace Bro2Bro.Mobile.Droid.Interfaces {    
     public class LocalStorage : ILocalStorage {
         public void Write<T>(string filename, T objectValue) {
             var documentsPath = Environment.GetFolderPath(Environment.SpecialFolder.Personal);
@@ -20,22 +25,26 @@ namespace Bro2Bro.Mobile.Droid.Interfaces {
             File.WriteAllBytes(filePath, fileBytes);
         }
 
-        public T Read<T>(string filename) {
+        public ReturnSet<T> Read<T>(string filename) {
             var documentsPath = Environment.GetFolderPath(Environment.SpecialFolder.Personal);
             var filePath = Path.Combine(documentsPath, filename);
-            
+
+            if (!File.Exists(filePath)) {
+                return new ReturnSet<T>($"{filePath} does not exist");
+            }
+
             var fileBytes = File.ReadAllBytes(filePath);
 
             var jsonStr = System.Text.Encoding.UTF8.GetString(fileBytes);
 
-            return (T)JsonConvert.DeserializeObject(jsonStr);
+            return new ReturnSet<T>((T)JsonConvert.DeserializeObject(jsonStr));
         }
 
         public void Write<T>(T objectValue) {
             Write(objectValue.GetType().Name, objectValue);
         }
 
-        public T Read<T>(Type objectType) => Read<T>(objectType.Name);
+        public ReturnSet<T> Read<T>(Type objectType) => Read<T>(objectType.Name);
 
         public void Delete<T>(T objectValue) {
             Delete<T>(objectValue.GetType().Name);
@@ -44,6 +53,10 @@ namespace Bro2Bro.Mobile.Droid.Interfaces {
         public void Delete<T>(string fileName) {
             var documentsPath = Environment.GetFolderPath(Environment.SpecialFolder.Personal);
             var filePath = Path.Combine(documentsPath, fileName);
+
+            if (!File.Exists(filePath)) {
+                return;
+            }
 
             File.Delete(filePath);
         }
